Debugging now...

+++ src/Control/li_ctrl_mode.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-enum class CTRL_MODE_LIST : int
+enum class CTRL_MODE_LIST : uint16_t
 {
     STAY,
     POSITION,
